+++ package.json
@@ -2,9 +2,9 @@
   "name": "blog",
   "version": "1.0.0",
   "description": "Create blog using NodeJS base guide F8",
-  "main": "index.js",
+  "main": "src\\index.js",
   "scripts": {
-    "start": "nodemon --inspect index.js",
+    "start": "nodemon --inspect src\\index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -18,7 +18,8 @@
   },
   "homepage": "https://github.com/yugiking0/blog-f8#readme",
   "dependencies": {
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "express-handlebars": "^7.1.2"
   },
   "devDependencies": {
     "morgan": "^1.10.0",
