+++ src/resources/views/layouts/main.hbs
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Document</title>
+    </head>
+    <body>
+        <div class="app">            
+            {{> header}}
+
+            <div class="container">
+                {{{body}}}
+            </div>
+
+            {{> header}}
+        </div>
+    </body>
+</html>
