+++ src/resources/views/news.hbs
@@ -1,0 +1,3 @@
+<h1 style="color: red;">
+  Tin Tuc
+</h1>
