+++ src/resources/views/partials/footer.hbs
@@ -1,0 +1,1 @@
+<h1>FOOTER</h1>
