+++ src/resources/views/partials/header.hbs
@@ -1,0 +1,1 @@
+<h1>HEADER</h1>
